requirements.txt
websockets
uvicorn
python-dotenv
opencv-python-headless
fastapi
redis
google-cloud-aiplatform
google-genai
